Improved cohesion of s-attribute computation rules
Changed how actions rules are applied for s-attributes. Assignments of envirionment parameters are done by default and attribute logic operations are contained in the grammar.
